+++ app/src/scenes/suivi/index.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { SafeAreaView, StyleSheet, View, Dimensions, TouchableOpacity } from "react-native";
+import { SafeAreaView, StyleSheet, View } from "react-native";
 import { useFocusEffect } from "@react-navigation/native";
 
 import { beforeToday } from "../../utils/date/helpers";
@@ -7,29 +7,20 @@ import Header from "../../components/Header";
 // import ChartPicker from "./chartPicker";
 import ChartPicker from "./chartPicker2";
 import RangeDate from "./RangeDate";
-import ScorePicker from "./ScorePicker";
-import FriseScreen from "./frise/FriseScreen";
 import ChartPie from "./chartPie";
 import Evenements from "./events";
 import Courbes from "../calendar/calendar";
 import logEvents from "../../services/logEvents";
-import { colors } from "../../utils/colors";
-import Icon from "../../components/Icon";
 import localStorage from "../../utils/localStorage";
 import FloatingPlusButton from "../../components/FloatingPlusButton";
-import { FriseInfoButton } from "./frise/FriseInfoButton";
-import { Button2 } from "../../components/Button2";
+import { FriseScreen } from "./frise/FriseScreen";
 
-const screenHeight = Dimensions.get("window").height;
-
-const Suivi = ({ navigation, setPlusVisible, startSurvey }) => {
+const Suivi = ({ navigation, startSurvey }) => {
   const [chartType, setChartType] = React.useState("Frises");
+  const [presetDate, setPresetDate] = React.useState("lastDays7");
   const [fromDate, setFromDate] = React.useState(beforeToday(30));
   const [toDate, setToDate] = React.useState(beforeToday(0));
-  const [focusedScores, setFocusedScores] = React.useState([1, 2, 3, 4, 5]);
-  const [showTraitement, setShowTraitement] = React.useState(true);
   const [aUnTraiement, setAUnTraitement] = React.useState(false);
-  const [showHint, setShowHint] = React.useState(false);
 
   useFocusEffect(
     React.useCallback(() => {
@@ -65,18 +56,13 @@ const Suivi = ({ navigation, setPlusVisible, startSurvey }) => {
         return (
           <FriseScreen
             navigation={navigation}
+            presetDate={presetDate}
+            setPresetDate={setPresetDate}
             fromDate={fromDate}
+            setFromDate={setFromDate}
             toDate={toDate}
-            focusedScores={focusedScores}
-            showTraitement={showTraitement}
-            showHint={showHint}
-            onCloseHint={() => {
-              setShowHint(false);
-              logEvents.logSuiviShowLegendeInformationPriseDeTraitement(0); // 0 = masquer, 1 = afficher
-            }}
-            setShowHint={setShowHint}
-            aUnTraiement={aUnTraiement}
-            setShowTraitement={setShowTraitement}
+            setToDate={setToDate}
+            hasTreatment={aUnTraiement}
           />
         );
     }
@@ -88,7 +74,7 @@ const Suivi = ({ navigation, setPlusVisible, startSurvey }) => {
         <View style={styles.headerContainerNavigation}>
           <Header title="Mes analyses" navigation={navigation} />
         </View>
-        <View style={styles.headerContainer}>
+        <View style={styles.tabContainer}>
           <ChartPicker
             // onAfterPress={() => setChartType(nextChartType(chartType))}
             // onBeforePress={() => setChartType(prevChartType(chartType))}
@@ -96,78 +82,20 @@ const Suivi = ({ navigation, setPlusVisible, startSurvey }) => {
             onChange={(e) => setChartType(e)}
             ongletActif={chartType}
           />
-          {chartType !== "Courbes" ? (
+        </View>
+        {chartType !== "Courbes" && chartType !== "Frises" && (
+          <View style={styles.headerContainer}>
             <RangeDate
+              presetValue={presetDate}
+              onChangePresetValue={setPresetDate}
               fromDate={fromDate}
               toDate={toDate}
               onChangeFromDate={setFromDate}
               onChangeToDate={setToDate}
-              withPreset={chartType === "Frises" || chartType === "Statistiques"}
-            >
-              {chartType === "Frises" && (
-                <>
-                  <Button2
-                    title="Filtrer"
-                    icon="TuneSvg"
-                    size="small"
-                    type="outline"
-                    containerStyle={{ marginRight: 8 }}
-                  />
-                  <FriseInfoButton
-                    navigation={navigation}
-                    hasTreatment={aUnTraiement}
-                    containerStyle={{ position: "absolute", right: 0 }}
-                  />
-                </>
-              )}
-            </RangeDate>
-          ) : null}
-          {chartType === "Frises" ? (
-            <View style={styles.containerScorePickerFrise}>
-              <ScorePicker
-                focusedScores={focusedScores}
-                onPress={(i) => {
-                  if (focusedScores.includes(i)) {
-                    setFocusedScores((e) => e.filter((x) => x !== i));
-                  } else {
-                    setFocusedScores((e) => [...e, i]);
-                  }
-                  //events
-                  logEvents.logSuiviEditScoreFrise(i);
-                }}
-              />
-              <View style={styles.verticalDivider} />
-              <View style={styles.hintContainer}>
-                <TouchableOpacity
-                  onPress={() => {
-                    if (aUnTraiement) {
-                      setShowTraitement((e) => !e);
-                      logEvents.logSuiviShowPriseDeTraitement(showTraitement ? 0 : 1); // 0 = masquer, 1 = afficher
-                    } else {
-                      setShowHint((e) => !e);
-                    }
-                  }}
-                >
-                  <View
-                    style={[
-                      styles.selectionContainer,
-                      !aUnTraiement && styles.noTraitementSelectionContainer,
-                      showTraitement && styles.activeSelectionContainer,
-                    ]}
-                  >
-                    <Icon
-                      icon="DrugsSvg"
-                      color={!aUnTraiement || showTraitement ? "#FFFFFF" : "#58C8D2"}
-                      width={20}
-                      height={20}
-                      styleContainer={styles.icon}
-                    />
-                  </View>
-                </TouchableOpacity>
-              </View>
-            </View>
-          ) : null}
-        </View>
+              withPreset={chartType === "Statistiques"}
+            />
+          </View>
+        )}
         {renderChart(chartType)}
       </SafeAreaView>
       <FloatingPlusButton shadow onPress={startSurvey} plusPosition={0} />
@@ -175,102 +103,28 @@ const Suivi = ({ navigation, setPlusVisible, startSurvey }) => {
   );
 };
 
-const styles = StyleSheet.create({
-  hintContainer: {
-    marginVertical: 10,
-    marginHorizontal: 10,
-    flexDirection: "row",
-    alignItems: "center",
-    justifyContent: "space-around",
-    display: "flex",
-  },
-  infoHintContainer: {
-    width: 20,
-    height: 20,
-    display: "flex",
-    justifyContent: "center",
-    alignItems: "center",
-    borderColor: colors.LIGHT_BLUE,
-    borderWidth: 1,
-    borderRadius: 15,
-    marginLeft: 10,
-  },
-  activeInfoHintContainer: {
-    backgroundColor: colors.LIGHT_BLUE,
-  },
-  infoHintText: {
-    color: colors.LIGHT_BLUE,
-  },
-  activeInfoHintText: {
-    color: "#FFFFFF",
-  },
-  icon: {
-    width: 30,
-    height: 30,
-  },
-  selectionContainer: {
-    padding: 4,
-    borderColor: "#DEF4F5",
-    borderWidth: 1,
-    borderRadius: 10,
-  },
-  activeSelectionContainer: {
-    backgroundColor: colors.LIGHT_BLUE,
-  },
-  noTraitementSelectionContainer: {
-    backgroundColor: "#E9E9E9",
-    borderColor: "#DADADA",
-  },
-  containerScorePickerFrise: {
-    display: "flex",
-    flexDirection: "row",
-    justifyContent: "space-between",
-    alignItems: "center",
-  },
+export const styles = StyleSheet.create({
   headerContainerNavigation: {
     padding: 5,
     paddingBottom: 0,
     backgroundColor: "#1FC6D5",
   },
+  tabContainer: {
+    backgroundColor: "#FFFFFF",
+    flexDirection: "column",
+    justifyContent: "center",
+    alignItems: "center",
+  },
   headerContainer: {
     paddingBottom: 15,
     backgroundColor: "#FFFFFF",
-  },
-  imageContainer: {
-    display: "flex",
+    flexDirection: "column",
+    justifyContent: "center",
     alignItems: "center",
-  },
-  image: {
-    height: screenHeight * 0.5,
-    resizeMode: "contain",
-  },
-  subtitle: {
-    flex: 1,
-    color: "#000",
-    fontSize: 15,
-    fontWeight: "normal",
-  },
-  subtitleContainer: {
-    display: "flex",
-    flexDirection: "row",
-    marginVertical: 10,
-  },
-  bold: {
-    fontWeight: "bold",
   },
   safe: {
     flex: 1,
     backgroundColor: "#1FC6D5",
-  },
-  title: {
-    fontWeight: "700",
-    fontSize: 22,
-  },
-  verticalDivider: {
-    height: "50%",
-    backgroundColor: "#E0E0E0",
-    width: 1,
-    alignSelf: "center",
   },
 });
